pyproject.toml
[project]
name = "radiobuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy==1.26.4",
    "openai==2.26.0",
    "piper-tts>=1.4.1",
    "python-dotenv==1.0.1",
    "rich==13.9.4",
    "sounddevice==0.4.6",
    "soundfile==0.12.1",
]
